Fresheats
This assignment was absolutely amazing as I've learned so much of HTML, and CSS. I was able to deliver on timelly manner and met all of the assigned requirements. The challenge of this project was one that contributed to my skill sets as Jr Front-end Developer. Please see the below link for my final product:
https://htmlpreview.github.io/?https://github.com/adenyasmina/fresheats/blob/master/index.html